dcs_memory/services/glm/requirements.txt
grpcio
grpcio-tools
protobuf
qdrant-client
asyncpg # For PostgreSQL external connector
PyYAML
pydantic-settings
aiosqlite
python-json-logger
pytest
pytest-asyncio
grpcio-testing
grpcio-health-checking
